Refactor user to be associated to the account instead of user associated to the account
Through discussion we determined that it made more sense to associate an account with the user model. This model currently contains user's profile, relationships, and the DMs and servers the user is part of. These will be refactored in a future PR. They will each be associated with the user's account once they have been refactor out.

Changes in this PR:
-move the user name and discimator to the account model
-user model will have a account property to associate it with the user's account
-user controller will use id of the account to look up the profile
-relationship controller will use id of the account to do associated actions
-update the auth path to be seperate from account routes such as create, update, delete
-create new route module for get, create, update, delete of account